requirements.txt
# Core ML / DL — wheels oficiales para Windows, sin compilación C++
torch>=2.1.0
torchvision>=0.16.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0

# ── Detección de rostros ──────────────────────────────────────────────────────
# facenet-pytorch: PyTorch puro → MTCNN (Etapa 1) + FaceNet embedder (alternativa)
# Wheel oficial en PyPI, sin C++ ni Cython.
facenet-pytorch>=2.5.3

# ── Embeddings faciales ───────────────────────────────────────────────────────
# deepface: Python puro, descarga pesos ArcFace pre-compilados (.h5) en el
# primer uso → sin compilación. Reemplaza a insightface (que requería MSVC).
deepface>=0.0.92
tf-keras>=2.16.0           # backend Keras que necesita deepface con TF>=2.16

# ── Liveness Detection ───────────────────────────────────────────────────────
# DenseNet201 se carga vía torchvision (incluido en torch arriba)

# Data & Numerics
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
requests>=2.31.0

# Evaluation & Metrics
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Security (cifrado de embeddings en reposo)
cryptography>=41.0.0

# Configuration & Utilities
PyYAML>=6.0
python-dotenv>=1.0.0
kagglehub>=0.3.0
tqdm>=4.66.0
loguru>=0.7.0
pywhatkit>=5.4

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: vector DB para búsqueda eficiente a escala (wheel puro C precompilado)
faiss-cpu>=1.7.4
